Made styling changes to whole app, added side bar filter menu
