+++ src/src-frontend/Components/Navbar/Menu/Menu.css
@@ -1,11 +1,31 @@
-.menu-container {
+/* .menu-container {
   color: black;
   text-decoration: none;
   font-weight: bold;
+} */
+
+.menu-list {
+  width: 50%;
+  position: fixed;
+  top: 0;
+  left: 0;
+  transition: 0.5s;
+  text-align: left;
+}
+
+.filter-box {
+  padding: 1rem 0px;
+  font-size: 1.3rem;
+  font-weight: 500;
+  display: block;
+  transition: 0.3s;
+  list-style: none;
+  border-bottom: 1px solid orange;
+  background-color: white;
 }
 
 .menu-header {
-  font-family: 'Muli', sans-serif;
+  font-family: "Muli", sans-serif;
   font-size: 20px;
   font-weight: bold;
   background: none;
