+++ src/src-frontend/Components/Navbar/Menu/Menu.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import "./Menu.css";
 import filterIcon from "../../../Images/filter-icon.svg";
 import Recipe from "../../Navbar/Recipe/Recipe.js";
-import {useIsMount} from '../../Searchbar/useIsMount.js'
+import { useIsMount } from "../../Searchbar/useIsMount.js";
 // import { Link } from "react-router-dom";
 
 export default function Menu() {
@@ -11,8 +11,6 @@ export default function Menu() {
   const [recipes, setRecipes] = useState([]);
   const isMount = useIsMount();
 
-
- 
   useEffect(() => {
     if (isMount) {
       console.log("First Render");
@@ -22,52 +20,79 @@ export default function Menu() {
   }, [query]);
 
   const getSearch = (e) => {
-    setQuery(e.target.value); 
-  }
+    setQuery(e.target.value);
+  };
 
   const getRecipes = async () => {
-    const response = await fetch(`https://www.themealdb.com/api/json/v1/1/filter.php?c=${query}`);
+    const response = await fetch(
+      `https://www.themealdb.com/api/json/v1/1/filter.php?c=${query}`
+    );
     const data = await response.json();
     setRecipes(data.meals);
     console.log(data.meals);
   };
 
+  function openNav() {
+    setMenuOpen(!menuOpen);
+  }
+
   return (
     <>
       {/* // <div className="menu-container">  */}
-      <button className="filter-button" onClick={() => setMenuOpen(!menuOpen)}>
+      <button className="filter-button" onClick={openNav}>
         <img src={filterIcon} alt="filter button" />
       </button>
       {menuOpen && (
         <>
-        <div className="menu-list">
-            <button className='filter-box' value='Seafood' onClick={getSearch}>Seafood</button>
-            <button className='filter-box' value='Beef' onClick={getSearch}>Beef</button>
-            <button className='filter-box' value='Chicken' onClick={getSearch}>Chicken</button>
-            <button className='filter-box' value='Dessert' onClick={getSearch}>Dessert</button>
-            <button className='filter-box' value='Lamb' onClick={getSearch}>Lamb</button>
-            <button className='filter-box' value='Pasta' onClick={getSearch}>Pasta</button>
-            <button className='filter-box' value='Pork' onClick={getSearch}>Pork</button>
-            <button className='filter-box' value='Side' onClick={getSearch}>Side</button>
-            <button className='filter-box' value='Starter' onClick={getSearch}>Starter</button>
-            <button className='filter-box' value='Vegan' onClick={getSearch}>Vegan</button>
-            <button className='filter-box' value='Vegetarian' onClick={getSearch}>Vegetarian</button>
-            <button className='filter-box' value='Breakfast' onClick={getSearch}>Breakfast</button>
-        </div>  
-       
-        </>      
-      )
-      }
+          <ul className="menu-list">
+            <li className="filter-box" value="Seafood" onClick={getSearch}>
+              Seafood
+            </li>
+            <li className="filter-box" value="Beef" onClick={getSearch}>
+              Beef
+            </li>
+            <li className="filter-box" value="Chicken" onClick={getSearch}>
+              Chicken
+            </li>
+            <li className="filter-box" value="Dessert" onClick={getSearch}>
+              Dessert
+            </li>
+            <li className="filter-box" value="Lamb" onClick={getSearch}>
+              Lamb
+            </li>
+            <li className="filter-box" value="Pasta" onClick={getSearch}>
+              Pasta
+            </li>
+            <li className="filter-box" value="Pork" onClick={getSearch}>
+              Pork
+            </li>
+            <li className="filter-box" value="Side" onClick={getSearch}>
+              Side
+            </li>
+            <li className="filter-box" value="Starter" onClick={getSearch}>
+              Starter
+            </li>
+            <li className="filter-box" value="Vegan" onClick={getSearch}>
+              Vegan
+            </li>
+            <li className="filter-box" value="Vegetarian" onClick={getSearch}>
+              Vegetarian
+            </li>
+            <li className="filter-box" value="Breakfast" onClick={getSearch}>
+              Breakfast
+            </li>
+          </ul>
+        </>
+      )}
       <div className="recipe-category-results">
-      {recipes.map((recipe) => (
-              <Recipe
-                key={recipe.idMeal}
-                id={recipe.idMeal}
-                title={recipe.strMeal}
-                image={recipe.strMealThumb}
-              />
-            ))
-        }
+        {recipes.map((recipe) => (
+          <Recipe
+            key={recipe.idMeal}
+            id={recipe.idMeal}
+            title={recipe.strMeal}
+            image={recipe.strMealThumb}
+          />
+        ))}
       </div>
     </>
   );
