--- src/src-frontend/Components/Navbar/Navbar.css
@@ -1,37 +1,0 @@
-.navbar {
-  background-color: white;
-  opacity: 0.95;
-  width: 100%;
-  top: 0;
-  left: 0;
-  right: 0;
-  height: 10em;
-  overflow: hidden;
-  display: grid;
-  grid-template-columns: 1fr 2fr;
-  z-index: 9999;
-  align-content: center;
-  justify-items: center;
-}
-
-.navbar-item {
-  color: black;
-  text-decoration: none;
-  font-weight: bold;
-  font-family: "Muli", sans-serif;
-  font-size: 20px;
-}
-
-.navbar-item.carrate-avatar {
-  width: 60px;
-  padding: 5px;
-  padding-top: 25px;
-  margin-left: 30px;
-}
-
-.navbar-item.title {
-  padding: 5px;
-  margin-top: 10%;
-  margin-right: 20%;
-  font-size: 75px;
-}
