+++ src/src-frontend/Components/Navbar/SideBarPractice/Navbar.css
@@ -1,0 +1,8 @@
+nav {
+  width: 100%;
+  height: 65px;
+  border-bottom: 2px solid white;
+  padding: 0 20px;
+  display: flex;
+  justify-content: space-between;
+}
