+++ src/src-frontend/Components/Navbar/SideBarPractice/Navbar.js
@@ -1,0 +1,12 @@
+import React from "react";
+import Filter from "./Filter.js";
+import "./Navbar.css";
+
+export default function Navbar() {
+  return (
+    <nav>
+      <div>Navbar</div>
+      <Filter />
+    </nav>
+  );
+}
