+++ src/src-frontend/Pages/Homepage/Homepage.css
@@ -52,6 +52,7 @@ h2 {
 
 .meet-team {
   background-color: #ff6600;
+  outline: none;
   font-family: "Carme", sans-serif;
   color: white;
   padding: 1rem 2rem;
