+++ src/src-frontend/Pages/Homepage/Homepage.js
@@ -5,7 +5,6 @@ import "./Homepage.css";
 import Searchbar from "../../Components/Searchbar/Searchbar.js";
 import InspireButton from "../../Components/InspireButton/InspireButton.js";
 import Hero from "../../Components/Hero/Hero.js";
-import Team from "../../Components/Team/Team.js";
 
 export default function Homepage() {
   const history = useHistory();
