+++ src/src-frontend/Pages/RandomRecipe/RandomRecipe.css
@@ -1,100 +1,102 @@
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@200;400;600&display=swap');
+@import url("https://fonts.googleapis.com/css2?family=Poppins:wght@200;400;600&display=swap");
 
 * {
-    box-sizing: border-box;
+  box-sizing: border-box;
 }
 
-.meal{
-    display:flex;
-    background-color:#ffeedc;
-    border-radius: 5px;
-    box-shadow: 0 14px 28px rgba(0,0,0,0.25), 0 10px 10px rgba(0,0,0,0.22);
+.meal {
+  display: flex;
+  background-color: #ffeedc;
+  border-radius: 5px;
+  box-shadow: 0 14px 28px rgba(0, 0, 0, 0.25), 0 10px 10px rgba(0, 0, 0, 0.22);
 }
 
-.meal-img{
-    flex:1;
-    overflow: hidden;
-    border-top-left-radius:5px;
-    border-bottom-left-radius:5px;
- }
-
-.meal-img img{
-    width: 100%;
-    height: 100%;
-    object-fit:cover;
+.meal-img {
+  flex: 1;
+  overflow: hidden;
+  border-top-left-radius: 5px;
+  border-bottom-left-radius: 5px;
 }
 
-.meal-details{
-    flex: 2;
-    padding:25px;
-    text-align:left;
-    font-family: 'Poppins', sans-serif;
-    position:relative;
+.meal-img img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.meal-details {
+  flex: 2;
+  padding: 25px;
+  text-align: left;
+  font-family: "Poppins", sans-serif;
+  position: relative;
 }
 
 .meal-title {
-    margin: 0 0 20px;
-    color:#ff6600;
+  margin: 0 0 20px;
+  color: #ff6600;
 }
 
-.meal-instructions{
-    font-size:1em;
+.meal-instructions {
+  font-size: 1em;
 }
 
-.meal-info{
-    list-style-type: none;
-    padding:0;
-    display:flex;
+.meal-info {
+  list-style-type: none;
+  padding: 0;
+  display: flex;
 }
 
-.meal-info li{
-    margin-right: 2em;
-    font-size:1em;
-    font-weight: 400;
-    margin-top:1em;
+.meal-info li {
+  margin-right: 2em;
+  font-size: 1em;
+  font-weight: 400;
+  margin-top: 1em;
 }
 
-.meal-info li strong{
-    color:#ff6600;
-    display:block;
-    font-size:1em;
+.meal-info li strong {
+  color: #ff6600;
+  display: block;
+  font-size: 1em;
 }
 
-.btn{
-    background-color:#ff6600;
-    padding:15px 25px;
-    font-size:1em;
-    border: none;
-    border-radius: 5px;
-    text-decoration:uppercase;
-    font-weight: 600;
-    cursor:pointer;
-    box-shadow: 0 14px 28px rgba(0,0,0,0.20), 0 8px 8px rgba(0,0,0,0.20);
+.btn {
+  background-color: #ff6600;
+  outline: none;
+  padding: 15px 25px;
+  font-size: 1em;
+  border: none;
+  border-radius: 5px;
+  text-decoration: uppercase;
+  font-weight: 600;
+  cursor: pointer;
+  box-shadow: 0 14px 28px rgba(0, 0, 0, 0.2), 0 8px 8px rgba(0, 0, 0, 0.2);
 }
 
-.btn a{
-    text-decoration:none;
-    color:white;
-    letter-spacing: 1px;
+.btn a {
+  text-decoration: none;
+  color: white;
+  letter-spacing: 1px;
 }
 
-.meal-details .btn{
-    position: absolute;
-    right: -12px;
-    bottom: -12px;
+.meal-details .btn {
+  position: absolute;
+  right: -12px;
+  bottom: -12px;
 }
 
-.btn:hover{
-    background-color:#ff9966;
+.btn:hover {
+  background-color: #ff9966;
 }
 
-.recipe-container{
-    margin: 0 auto;
-    max-width: 100%;
-    width: 800px;
-    padding:25px;
-  }
+.recipe-container {
+  margin: 0 auto;
+  max-width: 100%;
+  width: 800px;
+  padding: 25px;
+}
 
-.inspired-heading, h1{
-    color:#ff6600;
+.inspired-heading,
+h1 {
+  color: #ff6600;
 }
